requirements.txt
fastapi
uvicorn
requests
Pillow
RGBMatrixEmulator